+++ src/tools/index.ts
@@ -4,3 +4,4 @@ export * from "./rooms.js";
 export * from "./messaging.js";
 export * from "./transport.js";
 export * from "./admin.js";
+export * from "./scopes.js";